Lazy load console object
